+++ lib/code/FingerPrint.d.ts
@@ -1,6 +1,0 @@
-/**
- * This functions working
- * @Param {null}
- * @return {Promise<string>} - resolve(string)
- */
-export declare const getDeviceId: () => Promise<string>;